Implemented simple detection for direction change
Gestures are stored as a queue of dictionaries of id -> {direction since last registered point, or NONE if this id is new}. The queue is capped to a certain length (history) and will have old points dropped to allow new additions (though it does not now).

Additionally, waiting too long between producing data points causes a new gesture to be started.

Known issues: callback is called twice per timestamp (and per frame?), possibly from different threads or different objects sharing a reference, resulting in double frees when the array is appended to.

Directionality is unsolved -- because we are ignoring some datapoints a priori for being too frequent, we may have a situation in which both x and y directions have surpassed their threshold and we get a diagonal movement.