Fixed no-direction bug
Dirty fix: the touch-data pointer parameter was being reused by the callback, so storing the pointer was not enough to retain the previous frame. It now used memcpy to keep a copy of the whole input parameter (this should be fixed in a later general refactoring).

As a side effect, made gesture termination detection a bit more robust. The diagonalization problem is still a concern, however.