Directionality committed to compass directions
Diagonals are too fine a distinction and are not otherwise supported by the OS. There is enough variation in only the compass directions to create enough gestures. In the case where both x and y have varied enough to be considered a movement, the larger of the two dominates and is considered the only movement.

Stubbed implementation of distance function to help catch very fast gestures. This may not be necessary with a sufficiently high fixed sampling rate.