Fixed bug with missing fingers, disallowed changing fingers.
The range for identifiers was incorrectly [0, 10] rather than [1,

When the number of fingers changes during a gesture, that gesture
is submitted as is to detection and a gesture is begun with the
new set of fingers. While this disallowed tiptaps, etc., it makes
reasoning about swipes and all motions that don't require tapping
significantly easier.