Gestures are now canonicalized.
Touches are canonicalized as they are added: multiple consecutive
touches moving in the same direction are represented only by the
last of the consecutive touches (in other words, each piece of the
swipe for a single finger represented only by that piece's endpoint,
which also notes the direction of movement).

Velocity information is lost in this transformation, however, that
was not data that was being used to detect gestures.