Added gesture-start threshold.
In order to be considered nonstationary during the gesture, each
finger needs to move at least a certain distance (larger than the
minimum distance between sample points). This assists in increasing
accuracy of gestures where fingers are supposed to be stationary,
but actually move slightly due to error.

Hopefully (?) reduced the amount of work involved in resetting the
gesture after a failure.

Gestures that don't move anywhere (i.e. a tap without a swipe) are
no longer considered gestures.