Added sorting to gestures.
Fingers in a gesture are now sorted so that they can be compared to
other gestures. Each finger is sorted in ascending order based on the
distance from its starting point to the line at 135 degrees through
the origin. In this way, horizontally- and vertically-aligned fingers
can both be ordered (left-to-right and bottom-to-top, respectively)
and robustly.

Behavior on diagonally-aligned fingers is unhandled and likely
problematic, though given that gestures are only cardinal directions
at the moment, this is not expected to change.