Simple gesture detection!
For testing, the first gesture you make after startup is saved as
the target gesture, and any time you make this gesture again it
prints out to verify.

Removed -reset from Gesture and changed the listener to make a new
copy instead of resetting and reusing the same instance repeatedly.

Fixed bug in equality detection.